+++ core/org.eclipse.cdt.core/utils/org/eclipse/cdt/utils/som/parser/SomSymbol.java
@@ -11,7 +11,6 @@
 package org.eclipse.cdt.utils.som.parser;
 
 import java.io.IOException;
-import java.math.BigInteger;
 
 import org.eclipse.cdt.core.IAddress;
 import org.eclipse.cdt.utils.Addr2line;
@@ -59,7 +58,7 @@ public class SomSymbol extends Symbol {
 		Addr2line addr2line = ((SOMBinaryObject)binary).getAddr2line(true);
 		if (addr2line != null) {
 			try {
-				return addr2line.getLineNumber(getAddress().add(BigInteger.valueOf(offset)));
+				return addr2line.getLineNumber(getAddress().add(offset));
 			} catch (IOException e) {
 				// ignore
 			}
