+++ core/org.eclipse.cdt.core/utils/org/eclipse/cdt/utils/xcoff/parser/XCOFFBinaryObject.java
@@ -11,7 +11,6 @@ package org.eclipse.cdt.utils.xcoff.parser;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -206,7 +205,7 @@ public class XCOFFBinaryObject extends BinaryObjectAdapter {
 
 						IPath file = filename != null ? new Path(filename) : Path.EMPTY;
 						int startLine = addr2line.getLineNumber(addr);
-						int endLine = addr2line.getLineNumber(addr.add(BigInteger.valueOf(size - 1)));
+						int endLine = addr2line.getLineNumber(addr.add(size - 1));
 						list.add(new XCoffSymbol(this, name, type, addr, size, file, startLine, endLine));
 					} catch (IOException e) {
 						addr2line = null;
