+++ core/org.eclipse.cdt.core/utils/org/eclipse/cdt/utils/xcoff/parser/XCoffSymbol.java
@@ -7,7 +7,6 @@
 package org.eclipse.cdt.utils.xcoff.parser;
 
 import java.io.IOException;
-import java.math.BigInteger;
 
 import org.eclipse.cdt.core.IAddress;
 import org.eclipse.cdt.utils.Addr2line;
@@ -58,7 +57,7 @@ public class XCoffSymbol extends Symbol {
 		Addr2line addr2line = ((XCOFFBinaryObject)binary).getAddr2line(true);
 		if (addr2line != null) {
 			try {
-				return addr2line.getLineNumber(getAddress().add(BigInteger.valueOf(offset)));
+				return addr2line.getLineNumber(getAddress().add(offset));
 			} catch (IOException e) {
 				// ignore
 			}
